Allow suppressing static_var_with_dynamic_init (#6)
CUDA disallows initialization of shared variables with non-default constructors. However, CUDA has a `pipeline_shared_state` object that has a non-default constructor, and it seems to be a common CUDA pattern to do:
```
#pragma nv_diag_suppress static_var_with_dynamic_init
...
__shared__ cuda::pipeline_shared_state<cuda::thread_scope::thread_scope_block, STAGES / SEQ_UNROLL> pipeline_state;
```

However, clang always errors on this code and does not support disabling the error. This PR switches the error to a warning that is default to an error, this allows using `clang diagnostic push/pop` to disable the warning for parts of the code.

Signed-off-by: James Bartlett <james@gimletlabs.ai>